requirements.txt
pandas
openpyxl
tkinter
openai
python-dotenv
requests 